+++ common/src/main/java/io/github/mortuusars/exposure/item/CameraItemClientExtensions.java
@@ -34,11 +34,13 @@ public class CameraItemClientExtensions {
     }
 
     public static float itemPropertyFunction(ItemStack stack, ClientLevel clientLevel, LivingEntity livingEntity, int seed) {
-        if (stack.getItem() instanceof CameraItem cameraItem && cameraItem.isActive(stack)) {
-            if (cameraItem.isInSelfieMode(stack))
-                return livingEntity == Minecraft.getInstance().player ? 0.2f : 0.3f;
-            else
+        if (stack.getItem() instanceof CameraItem cameraItem) {
+            if (cameraItem.isActive(stack)) {
+                if (cameraItem.isInSelfieMode(stack))
+                    return livingEntity == Minecraft.getInstance().player ? 0.2f : 0.3f;
+
                 return 0.1f;
+            }
         }
 
         return 0f;